changed to float

+++ code_examples/BMICalc.py
@@ -1,9 +1,9 @@
 height = input("Input height (in cm): ")
 weight = input("Input weight (in kg): ")
 
-height = int(height)
+height = float(height)
 height = height / 100
-weightint = int(weight)
+weightint = float(weight)
 height2 = height ** 2
 bmi = weightint / height2
